==> 00_stim_prep/scripts_nb/01_Extract_CLIPtext.ipynb <==
from transformers import CLIPProcessor, CLIPModel
from tqdm.notebook import tqdm
import pickle
import torch
import numpy as np


def dump_data(data, filename):
    print('writing file: ' + filename)
    with open(filename, 'wb') as f:
        pickle.dump(data, f, pickle.HIGHEST_PROTOCOL)
        
captions = np.load("/projects/crunchie/boyanova/EEG_Things/eeg_prep/scripts/fmri_train_sim_caps.npy", allow_pickle=True)
model = CLIPModel.from_pretrained("openai/clip-vit-large-patch14")
processor = CLIPProcessor.from_pretrained("openai/clip-vit-large-patch14")

embeddings_1 = []  
for cap1, cap2 in tqdm(captions):
    inputs_1 = processor(text=cap1, return_tensors="pt", padding=True)
    
    with torch.no_grad():  
        embed_1 = model.get_text_features(**inputs_1)
    embeddings_1.append(embed_1.detach().cpu().numpy())

embeddings_1 = np.array(embeddings_1)



CLIP_txt = {"captions": captions[:,0],
           "embeddings": embeddings_1.reshape(8640, -1)}

import pickle
def dump_data(data, filename):
    print('writing file: ' + filename)
    with open(filename, 'wb') as f:
        pickle.dump(data, f, pickle.HIGHEST_PROTOCOL)
        
dump_data(CLIP_txt, "/projects/crunchie/boyanova/EEG_Things/eeg_prep/scripts/CLIP_txt_fmri_768.pickle")
