==> 00_stim_prep/scripts_nb/02_Extract_CLIPvis.ipynb <==
import torch
import pickle
from transformers import CLIPProcessor, CLIPModel
import numpy as np
from tqdm.notebook import tqdm
import os
from PIL import Image

def dump_data(data, filename):
    print('writing file: ' + filename)
    with open(filename, 'wb') as f:
        pickle.dump(data, f, pickle.HIGHEST_PROTOCOL)

# Load CLIP model and processor
project_dir = "/projects/crunchie/boyanova/EEG_Things/Grouping-Embeddings"
image_dir = "/projects/crunchie/boyanova/EEG_Things/data_set/Images"

cache_dir = os.path.join(project_dir, "models")
fmri_stim = np.load(os.path.join(project_dir, "files", "fmri_train_stim.npy"), allow_pickle=True)
imagePaths = []

for im in tqdm(fmri_stim):
    im_cat = im.split(".")[0][0:-4]
    imagePaths.append(os.path.join(image_dir, im_cat, im))

model = CLIPModel.from_pretrained("openai/clip-vit-large-patch14",         
                                  torch_dtype=torch.float16,
                                  device_map="auto",
                                  cache_dir=cache_dir)
processor = CLIPProcessor.from_pretrained("openai/clip-vit-large-patch14",                                                                             cache_dir=cache_dir)
model = model.to("cuda")

batch_size = 8 
all_image_embeddings = []

# Process images in batches
for i in tqdm(range(0, len(imagePaths), batch_size)):
    # Load and preprocess a batch of images
    batch_images = []
    for path in imagePaths[i:i + batch_size]:
        image = Image.open(path).convert("RGB")
        batch_images.append(image)

    # Use the CLIP processor to prepare the batch of images
    inputs = processor(images=batch_images, return_tensors="pt", padding=True).to("cuda")

    # Obtain the image embeddings
    with torch.no_grad():
        image_embeddings = model.get_image_features(**inputs)

    # Detach the embeddings and move to CPU
    all_image_embeddings.append(image_embeddings.detach().cpu().numpy())

# Convert the list of embeddings to a single NumPy array
all_image_embeddings = np.concatenate(all_image_embeddings, axis=0)
print("Total Image Embeddings Shape:", all_image_embeddings.shape)


CLIP_vis = {"stimuli": fmri_stim,
            "stimuli_paths": imagePaths,
            "embeddings": all_image_embeddings}

file_name = "CLIP_vis_fmri.pickle"
save_dir = os.path.join(project_dir, "files", file_name)
dump_data(CLIP_vis, save_dir)
